requirements.txt
aioredis==1.2.0
asgiref==3.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.6.2
Automat==0.7.0
certifi==2019.3.9
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==2.3.0
dj-database-url==0.5.0
Django==2.2.1
django-cors-headers==3.0.2
django-environ==0.4.5
django-heroku==0.3.1
django-monthfield==0.1.1
django-rest-swagger==2.2.0
djangorestframework==3.9.4
djangorestframework-camel-case==1.0.3
drf-nested-routers==0.91
entrypoints==0.3
flake8==3.7.7
geographiclib==1.49
geopy==1.20.0
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.1
openapi-codec==1.3.2
paho-mqtt==1.4.0
Pillow==6.0.0
psycopg2==2.8.2
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pytz==2019.1
redis==3.2.1
requests==2.22.0
sentry-sdk==0.9.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.1
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.3
uWSGI==2.0.18
whitenoise==4.1.2
zope.interface==4.6.0
